完成公告管理和管理员验证

+++ 源代码/admin/delCate.php
@@ -1,6 +1,8 @@
 <?php
 require_once '../config/config.php';
 require_once '../config/connect_db.php';
+require_once 'checkAdmin.php';
+checkAdmin();
 if(isset($_GET['id'])==TRUE){
     $catesql = "delete from category where id=".$_GET['id'];
     $cateres = $db->query($catesql);
